# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  entities.CheckPointRequest:
    properties:
      lat:
        maximum: 90
        minimum: -90
        type: number
      lng:
        maximum: 180
        minimum: -180
        type: number
    required:
    - lat
    - lng
    type: object
  entities.CheckPointResponse:
    properties:
      distance_km:
        type: number
      in_radius:
        type: boolean
      radius_km:
        type: number
    type: object
...
  entities.ForgotPasswordRequest:
    properties:
      email:
        type: string
    required:
    - email
    type: object
...
  entities.LoginRequest:
    properties:
      email:
        type: string
      password:
        type: string
    required:
    - email
    - password
    type: object
  entities.LoginResponse:
    properties:
      requires_2fa:
        type: boolean
      temp_token:
        type: string
    type: object
  entities.LogoutRequest:
    properties:
      refresh_token:
        type: string
    required:
    - refresh_token
    type: object
...
  entities.ProviderLocation:
    properties:
      delivery_radius_km:
        type: number
      id:
        type: string
      lat:
        type: number
      lng:
        type: number
      provider_id:
        type: string
      updated_at:
        type: string
    type: object
...
  entities.RefreshRequest:
    properties:
      refresh_token:
        type: string
    required:
    - refresh_token
    type: object
...
  entities.ResetPasswordRequest:
    properties:
      new_password:
        minLength: 8
        type: string
      token:
        type: string
    required:
    - new_password
    - token
    type: object
...
  entities.SetLocationRequest:
    properties:
      delivery_radius_km:
        maximum: 500
        type: number
      lat:
        maximum: 90
        minimum: -90
        type: number
      lng:
        maximum: 180
        minimum: -180
        type: number
    required:
    - delivery_radius_km
    - lat
    - lng
    type: object
...
  responses.APIResponse:
    properties:
      data: {}
      message:
        type: string
      success:
        type: boolean
    type: object
host: localhost:8080
info:
  contact:
    email: soporte@visionprice.app
    name: VisionPrice Support
  description: API backend para el módulo de Proveedores de VisionPrice
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: VisionPrice Proveedores API
  version: "1.0"
paths:
...
  /api/v1/auth/forgot-password:
    post:
      consumes:
      - application/json
      description: 'Envía un correo con un enlace para restablecer la contraseña (stub:
        se registra en logs)'
      parameters:
      - description: Correo electrónico
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/entities.ForgotPasswordRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/responses.APIResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/responses.APIResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/responses.APIResponse'
      summary: Solicitar recuperación de contraseña
      tags:
      - Auth
  /api/v1/auth/login:
    post:
      consumes:
      - application/json
      description: Autentica un proveedor con correo y contraseña. Retorna un token
        temporal para el flujo de 2FA.
      parameters:
      - description: Credenciales de acceso
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/entities.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/responses.APIResponse'
            - properties:
                data:
                  $ref: '#/definitions/entities.LoginResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/responses.APIResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/responses.APIResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/responses.APIResponse'
      summary: Iniciar sesión
      tags:
      - Auth
  /api/v1/auth/logout:
    post:
      consumes:
      - application/json
      description: Revoca el refresh token proporcionado para impedir su uso futuro
      parameters:
      - description: Refresh token a revocar
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/entities.LogoutRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/responses.APIResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/responses.APIResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/responses.APIResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/responses.APIResponse'
      security:
      - BearerAuth: []
      summary: Cerrar sesión
      tags:
      - Auth
  /api/v1/auth/refresh:
    post:
      consumes:
      - application/json
      description: Emite un nuevo access token usando un refresh token válido (verificando
        que no esté revocado)
      parameters:
      - description: Refresh token
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/entities.RefreshRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/responses.APIResponse'
            - properties:
                data:
                  additionalProperties:
                    type: string
                  type: object
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/responses.APIResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/responses.APIResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/responses.APIResponse'
      summary: Refrescar access token
      tags:
      - Auth
...
  /api/v1/auth/reset-password:
    post:
      consumes:
      - application/json
      description: Restablece la contraseña usando el token de recuperación
      parameters:
      - description: Token y nueva contraseña
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/entities.ResetPasswordRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/responses.APIResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/responses.APIResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/responses.APIResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/responses.APIResponse'
      summary: Restablecer contraseña
      tags:
      - Auth
...
  /api/v1/providers/location:
    get:
      description: Retorna la ubicación actual y radio de entrega del proveedor
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/responses.APIResponse'
            - properties:
                data:
                  $ref: '#/definitions/entities.ProviderLocation'
              type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/responses.APIResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/responses.APIResponse'
      security:
      - BearerAuth: []
      summary: Obtener ubicación del proveedor
      tags:
      - Geolocations
    put:
      consumes:
      - application/json
      description: Guarda la latitud, longitud y radio de entrega del establecimiento
      parameters:
      - description: Datos de ubicación
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/entities.SetLocationRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/responses.APIResponse'
            - properties:
                data:
                  $ref: '#/definitions/entities.ProviderLocation'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/responses.APIResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/responses.APIResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/responses.APIResponse'
      security:
      - BearerAuth: []
      summary: Fijar o actualizar ubicación del proveedor
      tags:
      - Geolocations
  /api/v1/providers/location/check:
    post:
      consumes:
      - application/json
      description: Calcula la distancia entre un punto dado y la ubicación del proveedor,
        y verifica si está dentro del radio
      parameters:
      - description: Coordenadas del punto a verificar
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/entities.CheckPointRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/responses.APIResponse'
            - properties:
                data:
                  $ref: '#/definitions/entities.CheckPointResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/responses.APIResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/responses.APIResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/responses.APIResponse'
      security:
      - BearerAuth: []
      summary: Verificar si un punto está dentro del radio de entrega
      tags:
      - Geolocations
securityDefinitions:
  BearerAuth:
    description: 'Ingrese el token JWT con el prefijo Bearer. Ejemplo: "Bearer {token}"'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"